+++ Changelog.md
@@ -6,6 +6,10 @@ Format for this file derived from [http://keepachangelog.com](http://keepachange
 
 ### Changed
 
+- *BREAKING*: Renamed `OGR::GeometryExtensions::EWKBIO` to `OGR::Geometry::EWKBIOExtensions`.
+- *BREAKING*: Renamed `OGR::GeometryExtensions::EWKBRecord` to `OGR::Geometry::EWKBRecord`.
+- *BREAKING*: Renamed `OGR::GeometryExtensions::LWGeomWrapper` to `OGR::Geometry::LWGeomExtensions`.
+- *BREAKING*: Renamed `OGR::GeometryExtensions::WKBRecord` to `OGR::Geometry::WKBRecord`.
 - *BREAKING*: All methods with default boolean args are now keyword args.
 - *BREAKING*: `required_ruby_version` set to `>= 2.6`